refactor: rename /members to /settings, fix breadcrumbs, fix cog links
- Move [handle]/members/page.tsx → [handle]/settings/page.tsx (replaces
  the old leave/delete-only settings page with the full config page)
- Breadcrumb: /settings now shows "Workspace settings" (no parent link)
  instead of the incorrect "Workflows / Members"
- Sidebar cog icon updated from /members → /settings
- Member avatars link updated from /members → /settings
- Profile header cog icon (near Workspace badge) already pointed to
  /settings — now correctly reaches the full config page after rename
